Rename Form1 to Login and update namespace
Renamed the form class from `Form1` to `Login` in `Form1.Designer.cs` and `Form1.cs`. Updated the `Name` and `Text` properties of the form to "Login". Modified the constructor name in `Form1.cs` to `Login`. Updated all references in the designer file to reflect the new class name. Added a `using` directive for `System.Net.Mime.MediaTypeNames` in the namespace declaration.